requirements.txt
asgiref==3.5.2
Django==3.2
django-currentuser==0.5.3
django-tinymce==3.5.0
Pillow==9.2.0
pkg_resources==0.0.0
pytz==2022.5
sqlparse==0.4.3
